Add ability to output all keys and values in thread context
The keys in a thread context can differ for different code paths. It
would be good if one could log whatever keys and values exist in the
thread context without specifying all possible keys in the logging
configuration.

This change modifies the "context" placeholder so that it can be used
without specifying a key. When no key is specified, all keys and their
values are sent to the output.