Make DynamicPathTest pass on any machine
On my machine, I see a unit test failure when DynamicPathTest runs. The
error message is

DynamicPathTest$GeneratingNewPath.generateDynamicPath:124
expected: /tmp/junit1072668570111541721/1970-01-01.log
 but was: /tmp/junit1072668570111541721/1969-12-31.log

The test generates a filename using a date string in the format
yyyy-MM-dd. The value for the date comes from a clock object set to
Instant.EPOCH, which is defined as midnight on 1970-01-01 in UTC. But
before being formatted for output, the date value is converted to a
ZonedDateTime. (See DynamicPath#generateNewPath.) The zone id is set to
the system default, which on my machine is "America/New_York." In that
time zone, Instant.EPOCH falls on 1969-12-31. The formatter thus outputs
"1969-12-31," which doesn't match the expected filename, "1970-01-01."
So the unit test fails.

This change explicitly sets the zone id used for the ZonedDateTime to
UTC instead of whatever the system default zone id is. This makes the
generated filename predictably come out as "1970-01-01," so the unit
test will always pass.