Address code review feedback
A previous commit introduced an ehancement allowing the user to output
all thread context keys and values by using the "context" placeholder
without specifying a key. This commit addresses feedback from a review
of that code. Notably, the ContextPlaceholder class has been refactored
into a class hierarchy: an abstract base class, a subclass that
encapsulates the logic of a context placeholder with a key, and a
subclass that encapsulates the logic of a context placeholder without a
key.